warning fixes

+++ Userland/SampleCodeModule/stdLib.c
@@ -106,7 +106,7 @@ void sysGetChar(char * ch){
 }
 
 void sysTestSuite(int num){
-    sysCall(6,num,0,0,0,0);
+    sysCall(6,(uint64_t)num,0,0,0,0);
 }
 
 void sysGetTime(int * buffer){
@@ -130,7 +130,7 @@ void sysPaintPixel(int x, int y, char B, char G, char R) {
 void sysExecute(void* functionPointer, char* name)
 {
   sysPrintString("Entered sysExecute\n",0,0,0);
-  sysCall(7,(uint64_t)functionPointer,name,0,0,0);
+  sysCall(7,(uint64_t)functionPointer,(uint64_t)name,0,0,0);
 }
 
 int getPid()
